Database-helsesjekk ligger i interfacet siden den er lik i begge implementasjonene.
